Encrypt communications between client and server
This is a fairly large patch, which introduces a "crypto" package
wrapping some simple AES encrypt/decrypt functions and a "config"
package that lets the client and the server retrieve their shared key
from a file on disk.

All commands sent from the server to the client are encrypted, and
encryption failures are logged in their output.